requirements.txt
clint
requests
pillow